refactor: BaseEntity import 방식 상대 경로로 수정
