+++ apps/backend/src/features/user/user.entity.ts
@@ -1,6 +1,6 @@
 import { USER_NICKNAME_MAX_LENGTH } from '@web24/shared';
-import { BaseIdCreatedUpdatedDeletedEntity } from 'src/common/entities/base-id-created-updated-deleted.entity';
 import { Column, Entity } from 'typeorm';
+import { BaseIdCreatedUpdatedDeletedEntity } from '../../common/entities/base.entity';
 
 @Entity({ name: 'users' })
 export class User extends BaseIdCreatedUpdatedDeletedEntity {
